requirements.txt
fastapi
uvicorn
langchain
langchain_community
langchainhub
langchain-openai
langchain-google-genai
chromadb
sentence-transformers
pydantic 
pytesseract
Pillow 
langchain-huggingface 
python-dotenv
python-multipart